refactor(QYOG-20): components 폴더 고도화
