+++ src/components/UI/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Created on Thu Nov 30 2023
+ *
+ * Copyright (c) 2023 Your Company
+ */
+
+export * from "./Button";
